+++ src/CTrue.FsConnect/SimProperty.cs
@@ -48,7 +48,7 @@ namespace CTrue.FsConnect
         public SimProperty(FsSimVar simVar, FsUnit unit, SIMCONNECT_DATATYPE dataType)
         {
             Name = FsSimVarFactory.GetSimVarCode(simVar);
-            Unit = UnitFactory.GetUnitName(unit);
+            Unit = FsUnitFactory.GetUnitName(unit);
             DataType = dataType;
         }
     }
